Update conformance tests for networks field to move to standard
