+++ conformance/base/baseline_admin_network_policy/standard-egress-inline-cidr-rules.yaml
@@ -1,0 +1,33 @@
+apiVersion: policy.networking.k8s.io/v1alpha1
+kind: BaselineAdminNetworkPolicy
+metadata:
+  name: default
+spec:
+  subject:
+    namespaces:
+      matchLabels:
+        kubernetes.io/metadata.name: network-policy-conformance-gryffindor
+  egress:
+  # CIDR rules that test allow to specific IPs is done within the test by updating the CR
+  #- name: "allow-egress-to-specific-podIPs"
+  #  action: "Allow"
+  #  to:
+  #  - networks:
+  #    - luna-lovegood-0.IP
+  #    - cedric-diggory-0.IP
+  - name: "allow-egress-to-slytherin"
+    action: "Allow"
+    to:
+    - pods:
+        namespaceSelector:
+          matchLabels:
+            conformance-house: slytherin
+        podSelector:
+          matchLabels:
+            conformance-house: slytherin
+  - name: "deny-egress-to-internet"
+    action: "Deny"
+    to:
+    - networks:
+        - 0.0.0.0/0
+        - ::/0
